+++ src/app/chatbot-design-studio/cds-dashboard/cds-canvas/actions/list/cds-actions.module.ts
@@ -40,8 +40,9 @@ import { CdsActionMakeComponent } from './cds-action-make/cds-action-make.compon
 import { CdsActionOnlineAgentsComponent } from './cds-action-online-agents/cds-action-online-agents.component';
 import { CdsActionOpenHoursComponent } from './cds-action-open-hours/cds-action-open-hours.component';
 import { CdsActionQaplaComponent } from './cds-action-qapla/cds-action-qapla.component';
-import { CdsActionReplaceBotV2Component } from './cds-action-replace-bot-v2/cds-action-replace-bot-v2.component';
-import { CdsActionReplaceBotComponent } from './cds-action-replace-bot/cds-action-replace-bot.component';
+import { CdsActionReplaceBotV2Component } from './cds-action-replace-bot/cds-action-replace-bot-v2/cds-action-replace-bot-v2.component';
+import { CdsActionReplaceBotV3Component } from './cds-action-replace-bot/cds-action-replace-bot-v3/cds-action-replace-bot-v3.component';
+import { CdsActionReplaceBotComponent } from './cds-action-replace-bot/cds-action-replace-bot-v1/cds-action-replace-bot.component';
 import { CdsActionReplyComponent } from './cds-action-reply/cds-action-reply-v1/cds-action-reply.component';
 import { CdsActionReplyButtonComponent } from './cds-action-reply/elements/cds-action-reply-button/cds-action-reply-button.component';
 import { CdsActionReplyFrameComponent } from './cds-action-reply/elements/cds-action-reply-frame/cds-action-reply-frame.component';
@@ -76,6 +77,8 @@ import { CdsActionClearTranscriptComponent } from './cds-action-clear-transcript
 import { CdsActionMoveUnassignedComponent } from './cds-action-move-unassigned/cds-action-move-unassigned.component';
 import { CdsActionConnectBlockComponent } from './cds-action-connect-block/cds-action-connect-block.component';
 import { CdsActionSendWhatsappComponent } from './cds-action-send-whatsapp/cds-action-send-whatsapp.component';
+import { VariableCssClassPipe } from 'src/app/pipe/variablecssClass.pipe';
+import { GetVariableNamePipe } from 'src/app/pipe/get-variable-name.pipe';
 
 @NgModule({
   declarations: [
@@ -103,6 +106,7 @@ import { CdsActionSendWhatsappComponent } from './cds-action-send-whatsapp/cds-a
     CdsActionDeleteVariableComponent,
     CdsActionReplaceBotComponent,
     CdsActionReplaceBotV2Component,
+    CdsActionReplaceBotV3Component,
     CdsActionAssignVariableComponent,
     CdsActionAssignVariableV2Component,
     CdsActionHideMessageComponent,
@@ -165,7 +169,8 @@ import { CdsActionSendWhatsappComponent } from './cds-action-send-whatsapp/cds-a
 
     //PIPES
     FilterPipe,
-    
+    GetVariableNamePipe,
+    VariableCssClassPipe
     
   ],
   imports: [
@@ -201,6 +206,7 @@ import { CdsActionSendWhatsappComponent } from './cds-action-send-whatsapp/cds-a
     CdsActionDeleteVariableComponent,
     CdsActionReplaceBotComponent,
     CdsActionReplaceBotV2Component,
+    CdsActionReplaceBotV3Component,
     CdsActionAssignVariableComponent,
     CdsActionAssignVariableV2Component,
     CdsActionHideMessageComponent,
@@ -259,7 +265,9 @@ import { CdsActionSendWhatsappComponent } from './cds-action-send-whatsapp/cds-a
     CdsActionSendWhatsappComponent,
     
     //PIPES
-    FilterPipe
+    FilterPipe,
+    GetVariableNamePipe,
+    VariableCssClassPipe
   ]
 })
 export class CdsActionsModule { }