+++ src/app/chatbot-design-studio/cds-rules/rules-add/action/action.component.ts
@@ -8,7 +8,7 @@ import { FormArray, FormGroup, FormGroupDirective } from '@angular/forms';
 })
 export class ActionComponent implements OnInit {
 
-  @Input() autocompleteOptions: string[]
+  @Input() autocompleteOptions: Array<{label: string, value: string}>;
   actionForm: FormGroup;
   constructor(private rootFormGroup: FormGroupDirective,) { }