+++ CHANGELOG.md
@@ -7,6 +7,60 @@
 ### **Copyrigth**: 
 *Tiledesk SRL*
 
+# 1.26.0
+
+# 1.26.0-rc.14
+- **added**: management for free expired plan
+
+# 1.26.0-rc.13
+- **bug-fixed**: copy and paste intent/action from different chatbots; 
+- **bug-fixed**: added vertical scroll to bot details page;
+
+# 1.26.0-rc.12
+- **added**: save the stage settings to local storage and load it at startup
+- **bug-fixed**: set the position to the top of the intent when the element height is greater than the stage size; 
+
+# 1.26.0-rc.11
+- **bug-fixed**: hide connector on mouseover only if alphaConnectors is 0
+
+# 1.26.0-rc.10
+- **bug-fixed**: panel-intent-header error intent_display_name
+- **added**: icon boy in panel-intent-detail
+- **added**: button to show/hide connectors
+
+# 1.26.0-rc.9
+- **bug-fixed**: autocompleteOption is not showed in action-replace-bot-v3
+
+# 1.26.0-rc.8
+- **added**: autocompleteOptions on blockName into action-replace-bot-v3
+- **changed**: autocompleteOptions in cds-text as array label-value
+
+# 1.26.0-rc.7
+- **changed**: botName with botId in action-replace-bot-v3
+
+# 1.26.0-rc.7
+- **changed**: botName with botId in action-replace-bot-v3
+
+# 1.26.0-rc.6
+- **added**: action-replace-bot-v3
+
+# 1.26.0-rc.5
+# 1.26.0-rc.4
+# 1.26.0-rc.3
+- **bug-fixed**: chatbot is not correctly selected in action-replace-bot-v2
+
+# 1.26.0-rc.2
+- **bug-fixed**: action-replace-bot-v2 not pass value correctly
+
+# 1.26.0-rc.1
+- **added**: slug as chatbot property and action-replace-bot-v2 option when user select a chatbot
+
+# 1.25.0-rc.3
+- **added**: intent and chatbot only agents visibility
+
+# 1.25.0-rc.2
+- **changed**: action-record translations and default values
+
 # 1.25.1
 
 # 1.25.0